backend-v1.1.7.1/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.6"
django-environ = "^0.11.2"
djangorestframework = "^3.14.0"
pillow = "^10.1.0"
drf-yasg = "^1.21.7"
drf-spectacular = "^0.26.5"
gunicorn = "^21.2.0"
python-dotenv = "^1.0.0"
psycopg2-binary = "^2.9.9"
drf-writable-nested = "^0.7.0"
celery = "^5.3.4"
redis = "^5.0.1"
django-celery-beat = "^2.5.0"
drf-chunked-upload = "^0.6.0"
django-typomatic = "^2.4.2"
django-filter = "^23.5"
django-cors-headers = "^4.3.1"
pytils = "^0.4.1"
boto3 = "^1.34.26"
django-storages = "^1.14.2"
django-ok-seo = "^0.9.4"
django-admin-sortable2 = "^2.1.10"
django-nested-admin = "^4.0.2"
django-tinymce = "^4.0.0"

[tool.poetry.dev-dependencies]
pre-commit = "^3.5.0"

[tool.poetry.group.dev.dependencies]
grpcio-tools = "^1.59.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.autoflake]
check = false
imports = ["fastapi", "pydantic"]

[tool.black]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 119
target-version = ['py39']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
skip_glob = ["migrations"]
skip = ["migrations"]
default_section = "THIRDPARTY"
sections = ["FUTURE", "STDLIB", "DJANGO", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]